+++ main.js
@@ -59,16 +59,16 @@ if (process.platform == 'darwin') {
 
 // FIXME: Garb. collector
 // Event for opening viewer
-ipcMain.on('open-viewer', (event, fileName, data) => {
+ipcMain.on('open-viewer', (event, fileName, mangaPath) => {
     nwin = new BrowserWindow({
         width: 700,
         height: 1080,
         //frame: false
     });
-    console.log("data is:" + data);
+    console.log("mangaPath is:" + mangaPath);
     nwin.loadURL(`file://${__dirname}/src/components/` + fileName + `.html`);
     nwin.webContents.on('dom-ready', function() {
-        nwin.webContents.send('send-to-viewer', data);
+        nwin.webContents.send('open-viewer-reply', mangaPath);
         nwin.show();
     });
 })
@@ -76,9 +76,9 @@ ipcMain.on('open-viewer', (event, fileName, data) => {
 ipcMain.on('open-chapter', (event, chapPath) => {
     console.log("chapPath is:" + chapPath);
     nwin.loadURL(`file://${__dirname}/src/components/viewer.html`);
-    mainWin.webContents.on('dom-ready', function() {
+    nwin.webContents.on('dom-ready', function() {
         event.sender.send('open-chap-reply', chapPath);
-        mainWin.show();
+        nwin.show();
     });
 })
