+++ src/components/viewer.html
@@ -21,6 +21,6 @@
     <script type="text/javascript" src="../../node_modules/jquery/dist/jquery.min.js"></script>
     <script type="text/javascript" src="../../node_modules/hammerjs/hammer.min.js"></script>
     <script type="text/javascript" src="../../node_modules/materialize-css/dist/js/materialize.js"></script>
-    <script type="text/javascript" src="../scripts/viewtab.js"></script>
+    <script type="text/javascript" src="../scripts/viewer.js"></script>
 </body>
 </html>
